Added support for updating diary entries.
Diary entries can now be updated by issuing PUT requests. Had to customize the mapping configuration to map symptoms properly as ModelMapper looks up the original target value via its accessor and we override them to rather return collection wrappers that it then tries to as collection, which breaks. This is now fixed by providing a custom Provider for the symptoms property that causes ModelMapper to use the instance produced by that rather than trying to look up the original collection.

Added shortcut on ErrorsDto to immediately produce a ResponseEntity from an Errors instances.