CORE-82 - Tightened checks of provided contact ways.
Introduced ContactWays value object to only accept if one of the contact ways (email, phone, mobile phone or identification hint) are provided. ContactPerson now requires an instance of that on construction to make sure at least one way is provided. Adapted DTO mapping and data initializers to use this new abstraction. The DTO now reports all four fields as invalid if none of them is set.