Rework of a lot of the security infrastructure.
Moved the remaining non-quarano packages into quarano.support and started to try to get rid of most of that code.

Introduced dedicated value objects for encrypted and unencrypted passwords to directly indicate which parts of the codebase see which variant of the password and parts that should only see an encrypted one cannot used with an unencrypted one by accident.

Refactored JWT Token creation and extracted a Spring Boot configuration properties class that now also contains most of the token creation and parsing code. The refactoring allowed to hide all of the implementation details and configuration and only expose a JwtTokenGenerator interface to allow the Account controller to generate a token on successful login.

Improve encapsulation in auth handling. Data initializer now uses the AccountService to not have to perform the password encoding itself. Simplified different reasons the Account creation can go wrong into a single exception type. AccountRegistrationDetails now lives in the auth module instead of the registration one. Added DTO validation to the submission of the account registration form. Also bind the repeated password and verify equality on submission.

Reimplemented ActivationCodeService based on Vavr's Try to reduce complexity in exception handling.