creating landing page for different user types
This landing page will be shown to the different user types when they
first visit the quarano app via the registration link of the
local health authority.
Depending on the type of user the text will differ and the call to action
button will lead the user to different entry points.
Furthermore, the registration code will be passed to the registration
page to pre-fill the form accordingly.