Improvements in diary entry handling.
We now do not render a "create" link for the evening slot of the first day if it's currently the morning slot. I.e. the newest entry that can be crated is the current slot.

Removed superfluous id field from DiaryEntryInput as it must never be submitted by the client in the first place.