Improved DiaryEntry timings.
We now again properly calculate all DiaryEntryDay instances. It was currently broken due to a bug in Slot causing the current day not being calculated at all.

We now also switched to 24h of post slot end for the editing cut off but allow retro-created entries to be edited within a certain tolerance configurable via quarano.diary.edit-tolerance defaulting to 15 minutes. It's explicitly configured to 2m only in the integrationtest profile so that we can easily play with this when running the app locally.