Make meta-annotations for general field validation work.
Tweaked the composed validation annotations to work based on the requirements of Bean Validation (we need a few superfluous attributes). Configured the annotation to be reported as single violation so that they can easily be i18ned via the simple annotation name in messages.properties.

Extended @Textual to allow more common characters of multi line text.