First draft of APIs for registration creation.
Tracked cases now expose a start-tracking link in case no account has yet been registered for the case. If a PUT request is issued to that link, an activation code will be generated.

Email templates are obtained from classpath resources located in masterdata/templates and need to be registered in the EmailTemplates bean defined in Quarano.