Model proper is new state for entities and aggregates.
We now implement Persistable with all entities and aggregates to properly expose whether an entity is considered new. This causes the creation date to be set properly now and the database communication significantly reduced as JPA doesn't have to constantly check for already existing instances during creation.

Reworked the Authentication implementation to always refer to a backing Account instance so that an AuditorAware implementation can access the account's identifier for auditing purposes.