Polish database model.
The primary goal of this change is to have readable and usable database model, that is accessible without JPA as well. Having the primary id of each table named id and a simple derived name from that one in the join tables as well as in the references makes joins much easier to write.

For the roles tables the primary id have been changed to an internal sequence instead of the hibernate one.

Obvious Copy&Paste "errors" in some of the identifiers have been fixed as well (i.e. comment should not have a primary id name "department_id“).

Therefor the commit includes the following changes to the underlying database structure:
- All primary id columns are named "id" (Some of those identifiers requires `AttributeOverride` as they are not relationships)
- All tables in plural names (there’s no hibernate strategy to do this, one could use a custom one plus the Inflector project, but that’s probably overkill)
- No reserved keywords (neither for columns nor for tables).