Polish database model and introduce Flyway and provide initial SQL scheme.
The primary goal of this change is to have readable and usable database model, that is accessible without JPA as well. Having the primary id of each table named id and a simple derived name from that one in the join tables as well as in the references makes joins much easier to write.

For the roles tables the primary id have been changed to an internal sequence instead of the hibernate one. Obvious Copy&Paste "errors" in some of the identifiers have been fixed as well (i.e. comment should not have a primary id name "department_id“).

Therefor the commit includes the following changes to the underlying database structure:

- All tables in plural names (there’s no hibernate strategy to do this, one could use a custom one plus the Inflector project, but that’s probably overkill)
- No reserved keywords (neither for columns nor for tables).

Introduced Flyway as a database migration tool and provides an initial scheme that is compatible with Hibernate JPA running against H2 and Postgres. To achieve this, the `BodyTemperature` class needs rework. It seems not possible to make either of those databases return a column definition that makes Hibernates validation happy.

However, as the temperature should be a decimal with at max a precision of 5 having a scale of 3, a Decimal(5,3) should work. The `BodyTemperature` domain type can encapsulate that change nicely.

Tweaks in schema generation.