Introduce Flyway and provide initial SQL scheme.
This commit introduces Flyway as a database migration tool and provides an initial scheme that is compatible with Hibernate JPA running against H2 and Postgres.
To achieve this, the `BodyTemperature` class needs rework. It seems not possible to make either of those databases return a column definition that makes Hibernates validation happy.

However, as the temperature should be a decimal with at max a precision of 5 having a scale of 3, a Decimal(5,3) should work. The `BodyTemperature` domain type can encapsulate that change nicely.