add ng serve configuration for ES5
As the local development configuration for ng serve only targets
ES2015 by default - which IE11 and older browsers do not support - we
had to add an additional configuration to compile the JS sources
in a format, that IE11 understands.
This is only needed for development as the production build uses
Differential Loading. This means the production build already generates
two kinds of bundles - one for browsers, that support ES2015 and newer
features and one for legacy browsers. So, the frontend code generated
via production build already works in IE11.

Added a new run script in package.json to run the frontend with ES5
configuration: `npm start-es5`.