Tweak database setup in various contexts.
There's now an AccountBootstrap that sets up a default department and account according to properties defined. See application.properties for examples. Unless a deployment environment overrides those, the default values will be used. The accounts will also only be created if no other department is already in place. The demo data creation trumps that.

Reinstantiated DatabaseInitializer to produce more output if it's getting active. Defensively skipping the wiping it performs if the database is not a Postgres database.

Changed roles to use UUIDs as well.