Fixed lookup of tracked cases without initial reports.
Fixed the mapping of the initial report field in TrackedCase as the nullable=false flag caused Hibernate to use a join that drops the entire result if no entry exists in the right hand side table. That caused cases without an initial report not being able to be looked up anymore.