CORE-149 - Fixes to SQL schema.
The reference to a tracked person within the encounters table cannot be non-null as Hibernate sets the id in a separate query.

Added test infrastructure that makes sure the JPA EntityManager gets flushed for each test execution to avoid changes not being written to the database before transaction rollback and thus the tests not really testing the database interaction.