CORE-189 Zweiter Nachtrag von Kontakten über Tagebuch funktioniert nicht
The TrackedPerson was created after a DiaryEntryAdded event but there
was no update event. Now an DiaryEntryUpdate event is registered for
changes to the entry and this event is handled.

Adds a test of DiaryEntry event handling, where first DiaryEntryAdded
and then DiaryEntryUpdated occurs.

Adds a check if there already a TrackedCase with the new contact person
in originContacts. If this check is missing, an existing contact of the
processed DiaryEntry is created again.