CORE-236 - Extract diary component from general tracking.
Extracted the diary handling code previously contained in the tracking one to be completely standalone. Tweaked client code to now consistently use the DiaryManagement instead of the repository.

Removed the obsolete handling of the DiaryUpdated event to trigger contact case creation as the introduction of the service now makes sure that updates *always* trigger the reporting of contacts and the handling of EncounterReported events is now sufficient. TrackedPerson.reportContact(…) now leniently handles calls reporting the same encounter by simply skipping duplicates and only publishing EncounterReported events in case of a new encounter. The event handling now also includes a simplification provided by @jekutzsche in #67 that simply checks for the presence of a tracked contact case that contains the currently handled contact person as original contact. That should also fix CORE-189.

The links pointing to the diary entry an encounter was originally created for are now provided by a RepresentationModelProcessor implemented in the diary package. This avoids the need to refer to diaries from the tracking package entirely. \ö/ A side effect of that change is that /api/encounters now returns a HAL collection rather than a top-level array which needs the frontend to adapt to rather lookup the encounters from $._embedded.encounters.

Introduced a MappingCustomizer interface to apply customizations to the ModelMapper instance we use as the type inspection taking place during the configuration considers fundamental customizations (e.g. the registration of type converters) already. Thus we need to make sure the order of customizations is stable which the previous arrangement of simply consuming the ObjectMapper in a bean did not provide as there's no stability guarantee in bean initialization ordering unless they somehow explicitly declare it or just depend on one another. Ported existing customizations to that new interface and apply those in the bean factory method providing the ModelMapper.

Related tickets: CORE-189.