CORE-222 - Fieber von Kontakt Fall nicht in Aktionsübersicht
The DiaryEntryActionItem was created after a DiaryEntryAdded event but
there was no update event. Now an DiaryEntryUpdate event is registered
for changes to the entry and this event is handled.

This handling is implemented for critical temperature and characteristic
symptoms.

Adds several tests for the DiaryEntry event handling for temperature and
symptoms.

Task-Url: https://jira.quarano.de/browse/CORE-222