CORE-258 - Creates contact cases for contacts of converted index cases.
When a contact case is converted to an index case, a new contact case is
created for all his contacts.

An event is used to separate changing the status of a case and
converting its contacts.

Adds an of-method to create a TrackedCase from a ContactPerson to avoid
code duplication.

https://quarano.atlassian.net/browse/CORE-258