CORE-119 - Adds API-GET method and representation for contacts of cases.
Adds the count of contacts to TrackedCaseDetails to enable the number to
be displayed.

Adds a link to the contacts representation of a TrackedCase to
TrackedCaseDetails.

Adds a repository method to find the TrackedCase that's associated with
a given ContactPerson.

Adds and implements an API-GET method to TrackedCaseController to get
the contacts representation of a TrackedCase. This returns a collection
of the new TrackedCaseContactSummary.

Story: Kontakte in der Falldetailseite anzeigen
URL: https://quarano.atlassian.net/browse/CORE-119