CORE-119 - Handles contact dates as list, extended the summary and refactoring.
The contact dates representation is now a list in one
TrackedCaseContactSummary for each contact person. This summary contains
a descending sorted list of all encounters. This attribute is called
descendingSortedContactDates to make it clearer how the list is sorted.

The new attributes caseId and caseType are necessary for routing in
frontend.

Adds attribute for contactStatus to the representation
TrackedCaseContactSummary.

The naming of the label attributes should make it clearer that this
are labels.

Code for generating the summaries was refactored and moved to the class
Encounters.

The code for determining the primary case type should be direct in the
enum.

Adds integration test for the use case and the REST-Interface.

Story: Kontakte in der Falldetailseite anzeigen
https://quarano.atlassian.net/browse/CORE-119