CORE-453: Execution of E2E Tests in GitHub
In general the run environment and the needed commands are nearly the same as in local dev env.
So they are triggered by simply running "npm run e2e".
To support that, we setup the environment with a german locale and timezone and start the backend
before running those e2e tests.
Afterwards the backend process is killed by selecting the correct process by its listening port 8080.
With that setup, we do NOT have to provide any CI-specific host variables, proxy settings, deployed backends or others.