CORE-252 - TrackedCases resource can now be filtered via q request parameter.
Added Querydsl JPA module to allow defining type safe predicates on repositories and integrated the APT plugin to build the meta-model properly. TrackedCaseRepository now exposes a method simply filter the tracked cases to be returned by a given String. For now, it is applied to the first- and last name of the TrackedPerson associated with the case. The controller takes that String via a "q" request parameter and forwards it to the repository.