CORE-252 - Apply and return origin cases for tracked cases.
The HTTP resource accepting tracked case details now also handles an array of URIs submitted via originCases and returns them in a HAL conforming $._embedded.originCases (see TrackedCaseDetails::new). The latter structurally confirms to the "select" representation introduced for tracked cases in general in previous commits by returning the name, birth date as well as a link pointing to the case. Those URIs have to be submitted with the PUT requests now so that they're kept around for updates.