CORE-279 - Fixed source case detection for vulnerable cases.
Tweaked the condition that eagerly aborts source case resolution as it was lacking detection of vulnerable cases. We now only abort if the source case is an index case in the first place.

Tweaked the test setup to use an @TestFactory to replicate the contact resolution test for all types of contact cases.