CORE-353 - Unified programming model for PUT and POST mappings in the controllers.
Refactored ErrorsDto into a more generic programming model for implementing handler methods that map PUT and POST requests. We now have MappedPayloads exposing factory methods to create a MappedError or MappedPayload type that allows monadic mapping over the payload, process the payload potentially accumulating further errors and producing a successful response on the happy path but automatically falling back to 400 and 404 if errors are accumulated or certain criterias are met.

Here's an example:

return MappedPayloads.of(payload, errors)
  .notFoundIf(…)                                         // always turn into 404 if predicate matches
  .alwaysMap((it, err) -> representations.from(it, err)) // invoke translation into domain object potentially producing additional errors
  .map(repository::save)                                 // invoke business operations on the entity
  .map(representations::toRepresentation)                // produce representation from result of the business operation
  .concludeIfValid(ResponseEntity::ok);                  // prepare response

MappedPayload allows to define predicates to eventually end up with a 404, invoke potentially error producing mapping steps, declare mapping steps to actually invoke business logic and produce a successful result if the processing did not produce any errors.

This required some changes in existing DTO classes that now rather return themselves than the Errors instance so that we can nicely ….map(…) over the invocations.

The i18n of errors is now postponed into an HttpMessageConverter which allows us to keep the controllers free of MessageSourceAccessors that previously had to be used to create the ErrorsDto instance in the first place.

Introduced …Representations classes in modules that so far didn't have those and moved all DTO types into those to avoid clutter.

The entire idea would benefit from better integration with Spring MVC's argument resolution but the APIs to implement that currently don't provide suitable access to neither the bound request body nor to the Errors produced during validation. I've filed a ticket [0] to improve that (hopefully for Spring Framework 5.3).

[0] https://github.com/spring-projects/spring-framework/issues/25411