CORE-231 - Polishing.
Refactorings in AccountService to avoid multiple paths through the service for password change efforts. All changes are now routed through the changePassword(…) call and whether to add an expiry date to the is based on the currently logged in user in relation to the account the new password is supposed to be set. The password is only encrypted if the current user is an administrator and she's not changing her own password in the first place. The latter guard is need to make sure an admin can create a non-expiring password for themselves. Also, EncryptedPassword is now immutable again.

As a side effect, the test data initializers don't need to set the password again to resolve the expiration, but the test case verifying the JSON representation of an account containing links to indicate the expired state now changes the password to an expiring one now.

Added another change-password link to the response of the authenticating call to make sure the client can detect that particular state and act accordingly. With a next link only, it wouldn't really know what to do as it's currently not interacting with the server to access the form. The combination of both links indicate what is to be done and that the client should prevent the user from doing anything but what's indicated.

Tweaked RoleRepository to assume that a role lookup by RoleType will always yield a non-null role. While null can technically be returned if no role is initially created for a RoleType, both the data initializer and AccountBootstrap actually take care of that. A non-nullable role lookup simplifies client code significantly.

Renamed AccountServiceMouleTests to AccountServiceIntegrationTests and PasswordTest to PasswordUnitTests.

Original pull request: #93.