CORE-345 - Automatic contact case creation now populates origin cases.
We now automatically populate the origin cases of a newly created contact case just as we do with the original contact information. To achieve this the concept of a case source that can either be constructed from a ContactPerson and TrackedCase or the latter only which would then create CaseSources for all registered encountered and their unique contacts.

Related tickets: CORE-252.