CORE-384 - Improved infrastructure for API and reference documentation.
Extended reference documentation to now contain first parts of the API documentation generated through Spring REST Docs. The document is now structured into API documentation first and backend developer documentation first.

To create the necessary artifacts for the former, we now leverage API introduced in src/docs/java which can be used for tests that solely serve the purpose of documenting particular flows. That said, the API (AbstractDocumentation, DocumentationFlow etc.) can also be used from already existing tests cases if they produce the flows we'd like to document.

The documentation of the payloads is implemented using the Spring Auto RESTDocs extension that automatically inspects the DTO classes for JSR-303 annotations. Also, it scans the Javadoc of those classes and uses that in the generated documentation. I.e. to document properties of the DTOs, simply write Javadoc for the getter methods.

Included Spring documentation resources for a more modern look and feel for the reference documentation. Included developing docs (previously etc/developing.adoc) into the reference documentation.