CORE-61 - implements automatic mail to new contact cases
- Adds dependencies to Spring Mail and GreenMail.
- Adds mail template for new contact mails.
- @EnableScheduling is now also active for profile integrationtest. The
existing Controller with scheduling have @Profile("!integrationtest")
now.
- Implements a generic EmailSender that determines the mail data from
and to from a given TrackedCase and sends a mail with the also given
subject and template as body.
- Implements a EmailServerMock that will started automatic in
integrationtests and is a buildin SMPTS and IMAPS server to handle
sended mails during development.
- Extends the TrackedCase and the database table to handle a mail
sending state.
- Implements a SSLSocketFactory to check the certificate of GreenMail.
- Implements a processor as an event listener of CaseCreated events that
sends automatic information mails about the contact to a COVID-19 case
to the person of the new contact case.
- The processor implements also ApplicationRunner to run after
DataInizialiser and avoid the sending of mails to automatic created
contact during development.
- The processor uses the generall EmailSender of Quarano and a template
for the mail body.
- Implements a new test class to test the processor function with a
GreenMail SMTP server mock.