CORE-357 - Tracked case details now also expose number of open anomalies.
We now allow the post-processing of a TrackedCaseDetails instance through TrackedCaseDetailsEnricher. The anomalies module registers one to enrich representations with both the number of open anomalies and also adds a link to point to the anomalies for that particular case. Also, the anomalies details point back to the original case now.