CORE-61 - Polishing.
Rearranged general email template handling by abstracting an interface the call to resolve those in an interface and tweaking the implementation to use the template's key to resolve the text file eventually. Turned the implementation package private by providing a custom configuration class within the core package to perform the actual setup and expose the bean. The GreenMail runtime dependency is now only added for the inmemory Maven profile to make sure it doesn't accidentally leak into a production deployment. By default it's only added in test scope. The configuration to use the dedicated SocketFactory to validate the GreenMail certificate has been moved to the application properties for the integration test profile.

EmailSender has been refactored to move more of the composition of a SimpleMailMessage into the AbstractTemplatedEmail, so that the primary interface TemplatedEmail only exposes a method to eventually compose the MailMessage instance. For ATE, two implementations based on a TrackedPerson / Department and TrackedCase exist from which the sender and recipient of the email is derived. Content of the email is selected by providing a Key to identify a template (see ConfigurableEmailTemplates and CoreConfiguration.emailTemplates(…)) alongside values for placeholders.

Folded the new mail for new contact case behavior into the already existing TrackedCaseEventListener.

Additional polish for nullability.