CORE-385 - Enable transaction management on package protected methods.
The default AnnotationTransactionAttributeSource registered by Spring Framework is configured to *only* consider public methods for transaction attributes. This probably stems from the time of JDK proxies in which all invocable methods on the proxy are public by definition. In the age of CGLib proxies however, package protected methods can be invoked on the proxy and pass it properly. Due to the default setting within AnnotationTransactionAttributeSource, the interceptor chain ends up empty as it doesn't even bother to inspect methods for transaction annotations.

We introduce a workaround in the form of a PriorityOrdered BeanPostProcessor that flips the flag within AnnotationTransactionAttributeSource to false, so that the lookup of the transaction attribute also inspects package protected methods.

Aside: tweaked JpaAutoFlushTestExceutionCallback to only flush if a transactions is actually running, as that is the scenario it is supposed to handle: to make sure changes can be written to the database even in the light of a following rollback, in which the persistence provider would skip the writing entirely. We need this additional tweak as we need to run the test for the changes described above in a non-transactional integration test and trying to flush the EntityManager without a running transaction fails.