CORE-373 + CORE-374 determines the language to be used and adds Content-Language Header
- Extends the table tracked_people with a column for the locale.
- Also the TrackedPerson and the TrackedPersonDataInitializer are
extended to handle the language of a user and initialize test data.
- The new LocaleResolver in LocaleConfiguration resolves the locale for
the request.
    1. Locale of the user
    2. Locale from Accept-Language-Header of the request
    3. de-DE as default
- Adds Content-Language Header with the used Language to every response.
This is done in a filter which is specified in the LocaleConfiguration.
- The @Bean-method messageSourceAccessor is moved to LocaleConfiguration
to collect als locale config in LocaleConfiguration.
- en and tr are supported at the moment
- Adds a message.properties file for english for tests.
- Adds tests for locale resolving and Content-Language setting.
- The login method of AuthenticationController sets an Authentication
after an successfull login. This is necessary so that LocaleResolver can
access the language of the current user at the end of the login request.