CORE-356 - Verifies the validity of the zip code and whether it belongs to.
the administered health department during registration

On PUT to /api/enrollment/details the zip code is checked with the RKI
data from HealthDepartments now. If no department exists a field error
with 400 will return. If the department for the zip code is unsupported,
the response is 422 with a message und the responsible department.

On PUT to /api/enrollment/details with parameter confirmed=true an
unsupported zip code is accepted. The case is then updated and marked as
external. The response can be 200 with the responsible department.

Parts of implementation:
- Extends Department and database by the RKI code. The example data
(Initializer) and tests are also adapted and extended.
- Adds a findByRkiCode method to DepartmentRepository for search with
RKI code.
- Extends status of a TrackedCase with EXTERNAL_ZIP to mark a case as an
external case.

- Moves code for RKI health departments from own module to department
module and delets the health_department module.

- Implements test for wrong, unsupported und confirmed unsupported zip
code in EnrollmentWebIntegrationTests.

CORE-356 Adapts handling of unsupported zip code to improved APIs to
create error responses.

Ollie has improved APIs to create error responses with commit
710f4c2aabfba891c69e4a8eb4bd6b782dda1c5a.