CORE-460 - Updates data from RKI and adapts code to new data.
The attributes Covid19Hotline, Covid19Fax and Covid19EMail are now also
read from the XML.

The DTO for a RKI HealthDepartment takes the covid19Hotline as phone
now, if this is available. Otherwise it uses the phone. A similar
procedure is used for e-mail and fax.

Adds a script to download the data from RKI and shortens the data by the
currently not needed place names.

The test are adapted to the changed data.