CORE-419 implements I18N for sympthoms
Extends the I18nedMessage with the defaultMessage, which is included in
MessageSourceResolvable. This can be used to set a static String that is
used when no code can be resolved.

Adds the symptom translation to the message….properties. Turkish is from
Google Translate!

The SymptomDto uses I18nedMessage for name now. So the ID of the Symptom
is used to resolve the translated name. The name from the database is
used as defaultMessage.