CORE-419 - I18N for symptoms.
Extends the I18nedMessage with the defaultMessage, which is included in
MessageSourceResolvable. This can be used to set a static String that is
used when no code can be resolved.

Adds the symptom translation to the message….properties.

The SymptomDto uses I18nedMessage for name now. So the ID of the Symptom
is used to resolve the translated name. The name from the database is
used as defaultMessage.

Original pull requests: #124, #134.