CORE-471 - Avoid repeated database lookups for i18n.
During message resolution, Spring calls LocaleResolver.resolveLocale(HttpServletRequest) for each message to be resolved. So far we've been looking up the preferred Locale to be used from the TrackedPerson that is tied to the current account. This commit introduces a ConcurrentHashMap based cache for the calculated Locale keyed by AccountIdentifier. That cache is potentially wiped every 60 seconds in case it exceeds 100 elements.

Once we upgrade to Spring Boot 2.4, we should switch to ConcurrentLruCache introduced in Spring Framework 5.3 to avoid the manual wiping.