CORE-494 - Rearrangement of Spring application profiles.
The primary Spring profile for local development is now "develop". It's still activated via a Maven profile (now named localDev) and during integration tests. As "develop" already existed, the semantics of that profile have been moved into a "staging" profile. That has two primary consequences:

1. IDE launch configurations for the application have to now use the "develop" profile to e.g. run against an in-memory H2 database.
2. The staging development deployment has to activate the "staging" profile instead of the "develop" one to trigger the database reset on deployment.

The integration test profile is still activated for the test annotations and used to reconfigure the log levels to warn (via a copy in src/test/resources). There's also a copy of the properties file for the tracing profile to avoid the log levels pulled up. Empty properties files for staging is kept around to make obvious the profile exists. The integration test one is only kept in test sources.