CORE-469 - Triggering start of registration process on a tracked case is now idempotent.
While we correctly re-set up a new activation code on a reinitiation of the case registration process, moving the case itself into the "in registration" state failed as it strictly expected the case to be in "open" state for that. In case a registration had been started before, it already was in the state "in registration" and thus rejected the state transition.

We now accept the state transition if the case already is in this state and effectively treat it as no-op.