CORE-463 - Department specific texts are now taken from the database.
The frontend and email texts are imported into the database over Flyway. There are common texts in scripts ending on .0 and specific texts in scripts ending on .1 under the RKI code as folder hierarchy.

For email templates/texts the old template system is used and a new DatabasedEmailTemplates as implementation of EmailTemplates is implemented.

For frontend texts the FrontendTextController uses the FrontendTextRepository to load the texts and returns this to the caller.

An I18nProperties is integrated to keep the default language central and to avoid errors with different specifications in the EmailTemplates and the LocaleConfiguration.

The old email template files ware deleted.

Original pull request: #145.