CORE-617 + CORE-620 Expose endpoints to export CSV data of cases in different formats
To the already existing endpoint for quarantine orders, there is now one
that creates the CSV for certain passed case IDs.

There are endpoints to export cases in the import formats (index or
contact) from Sormas and in an own Quarano format to transfer data to
other instances or to create statistics from it, for example. In the
latter case also for certain passed case IDs.

Specifically for transferring external cases to the appropriate health
departments, there is an endpoint that wraps one CSV for each department
in a JSON.

Another endpoint returns a sample CSV in Quarano's own format, which can
be used as a template for later imports.

To have the necessary information for the SORMAS import, Federal State
and District were added to the Department class. The service
"FederalStates" was added to determine the Federal State from the RKI
code and thus reduce the configuration.

The basis of exports is the TrackedCase class and its references. The
relevant data are transferred to a flat DTO TrackedCaseCsv and exported
from there to CSV. To control the different formats, custom
implementations of MappingStrategy and annotations related to them are
used.