CORE-487 - Documentation polish.
DocumentationFlow can now take custom response processors. We use that to avoid rendering the full HTML into the response for the documentation case for exposing frontend text.

Tweaked the table snippets for auto-generated docs to leave more space for the columns that carry verbose text.

Moved documentation of frontend texts to the back of the API docs as it's not a core API feature.

Original pull request: #143.