CORE-579 Prevents the sending of e-mails to any address during tests.
In Profile develop (here a local GreenMail-Server is started and used)
and prod the recipient addresses are taken further as they are. In other
profiles only @quarano.de addresses are used unchanged. No e-mail will
be sent to all other addresses. These will be sent to
testmailbox@quarano.de instead.