CORE-585 - API to send out registration emails.
The response of a registration (e.g. of GET /hd/cases/{id}/registration) contains a link send-mail now. With this link the e-mail can be sent and only if this link is available.

With a POST-request on /hd/cases/{id}/registration/mail the e-mail can be sent if a case exists for the given ID, it has an e-mail address and a valid activation code that has not yet been sent. The response is the current state of the registration.

A new state in ActivationCode persists if it has already been sent.

Original pull request: #165.