CORE-92 - Introduce API to reset a user's password.
A password reset can now be requested on an Account. It needs an email address and a username uniquely identifying the account. As the reset verification includes a check on a TrackedPerson's birthday, the final reset command has to be invoked on the newly introduced TrackedPersonManagement interface.

Sending out password reminder mails is triggered asynchronously via a composed annotation @AsyncTransactionalEventListener. Added reset password emails for German and English.

We now insist to receive an email address when creating an account.

Unified message keys for password errors under the Password.… prefix. The custom LocaleResolver now pulls the default locales from I18nMessages. Added a custom HttpMessageConverter to render MessageSourceResolvables out of the box.

Fixed security configuration for H2 console.

Added an HTTP resource for our API root to point to the primary resources for login and password reset. DocumentationFlow now only adds the authentication token headers if a TestSecurityContext is present.

Created TestEmailServer abstraction to simplify integration tests that verify email sending.