CORE-612 A fixed e-mail recipient can be set for tests.
Whether an email is sent to its real recipient no longer depends on the
Spring profile, as prod did not reliably identify the environment. With
the property spring.mail.properties.fixRecipient it is now possible to
set a fixed recipient address to which e-mails are sent instead of to
the real recipients. This is of course for testing purposes. If the
property is not set or set empty, the real recipient is used.

Mail to @quarano.de are always sent to this given address.