CORE-621 - Minimally invasive solution for the translation of placeholders in email texts.
In e-mails we have planned partly and at the moment generally the text in several languages. Sometimes placeholders have to be replaced according to the language of the respective part.

With this minimally invasive solution the emails are no longer assembled as strings but as EmailText instances. This allows the replacement of the placeholders for each language version separately. Furthermore, functions (Function<Locale, Object>) can now be used as replacement that are executed with the locale of the email part and can thus translate the replacement appropriately.

Original pull request: #166.