CORE-633 - Allow restarting the application for end to end tests.
Introduced the restart actuator via Spring Cloud in the localDev Maven profile to enable to restart the app quickly. Authenticated clients can issue a POST request to /actuator/restart to trigger an application context restart. Also enabled the readiness and liveness probes so that a client can trigger the restart and then call /actuator/health/readiness until $.status is "UP" before proceeding to issue requests.

Unfortunately, the restart actuator does not properly work with DevTools in play [0]. That means the e2e tests will have to actually build and package the application and then run it with the develop profile enabled. Adapted the CI workflow for e2e tests accordingly.

Cleaned up the Maven profiles accordingly. The localDev profile now skips Asciidoctor and Javadoc generation in case tests are disabled to speed up the packaging for sole running purposes.

[0] https://github.com/spring-cloud/spring-cloud-commons/issues/102